+++ application/core/forcegraph/src/commonMain/kotlin/io/writeopia/forcegraph/Graph.kt
@@ -1,0 +1,6 @@
+package io.writeopia.forcegraph
+
+import io.writeopia.forcegraph.model.Link
+import io.writeopia.forcegraph.model.Node
+
+data class Graph<T>(val nodes: List<Node<T>> = emptyList(), val links: List<Link<T>> = emptyList())
