+++ application/core/forcegraph/src/commonMain/kotlin/io/writeopia/forcegraph/model/Node.kt
@@ -1,9 +1,10 @@
 package io.writeopia.forcegraph.model
 
-data class Node(
+data class Node<T>(
+    val data: T,
     val x: Float,
     val y: Float,
     val vx: Float = 0f,
     val vy: Float = 0f,
-    val isDragged: Boolean = false
+    val isDragged: Boolean = false,
 )
