+++ application/features/note_menu/src/commonMain/kotlin/io/writeopia/notemenu/data/usecase/NotesNavigationUseCase.kt
@@ -1,6 +1,6 @@
 package io.writeopia.notemenu.data.usecase
 
-import io.writeopia.notemenu.data.model.NotesNavigation
+import io.writeopia.common.utils.NotesNavigation
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
