+++ application/features/note_menu/src/commonMain/kotlin/io/writeopia/notemenu/di/NotesMenuInjection.kt
@@ -1,7 +1,7 @@
 package io.writeopia.notemenu.di
 
 import androidx.compose.runtime.Composable
-import io.writeopia.notemenu.data.model.NotesNavigation
+import io.writeopia.common.utils.NotesNavigation
 import io.writeopia.notemenu.viewmodel.ChooseNoteViewModel
 
 interface NotesMenuInjection {
