+++ application/features/note_menu/src/commonMain/kotlin/io/writeopia/notemenu/ui/screen/menu/NotesMenuScreen.kt
@@ -9,7 +9,7 @@ import androidx.navigation.NavController
 import io.writeopia.commonui.dtos.MenuItemUi
 import io.writeopia.controller.OllamaConfigController
 import io.writeopia.model.ColorThemeOption
-import io.writeopia.notemenu.data.model.NotesNavigation
+import io.writeopia.common.utils.NotesNavigation
 import io.writeopia.notemenu.viewmodel.ChooseNoteViewModel
 
 @OptIn(ExperimentalSharedTransitionApi::class)
