+++ application/features/note_menu/src/commonMain/kotlin/io/writeopia/notemenu/viewmodel/FolderStateController.kt
@@ -5,7 +5,7 @@ import io.writeopia.common.utils.icons.IconChange
 import io.writeopia.common.utils.anyNode
 import io.writeopia.commonui.dtos.MenuItemUi
 import io.writeopia.sdk.models.document.Folder
-import io.writeopia.notemenu.data.usecase.NotesUseCase
+import io.writeopia.core.folders.repository.NotesUseCase
 import io.writeopia.sdk.models.document.MenuItem
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
