+++ application/features/note_menu/src/jvmTest/kotlin/io/writeopia/notemenu/viewmodel/ChooseNoteViewModelTest.kt
@@ -4,7 +4,7 @@ import io.mockk.coEvery
 import io.mockk.mockk
 import io.writeopia.auth.core.manager.AuthManager
 import io.writeopia.core.configuration.repository.ConfigurationRepository
-import io.writeopia.notemenu.data.usecase.NotesUseCase
+import io.writeopia.core.folders.repository.NotesUseCase
 import io.writeopia.ui.keyboard.KeyboardEvent
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
